// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/linking-ayah-preflight-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LinkingDescriptionChangesDto } from '../models/linking-description-changes-dto';
import { LinkingDoorWordImpactDto } from '../models/linking-door-word-impact-dto';
import { LinkingOverlappingSourceDto } from '../models/linking-overlapping-source-dto';
import { LinkingWordChangesDto } from '../models/linking-word-changes-dto';
export interface LinkingAyahPreflightDto {
  ayahId: number;
  ayahNumber: number;
  classification: string;
  descriptionChanges: LinkingDescriptionChangesDto;
  doorWordImpact: LinkingDoorWordImpactDto;
  invalidReason: string | null;
  overlappingSources: Array<LinkingOverlappingSourceDto>;
  surahNumber: number;
  verseKey: string;
  wordChanges: LinkingWordChangesDto;
}